Work without Metamask
Client should be able to work without meta-mask, using "ephemeral" key,
saved in a cookie.
This means that opening a new browser session will have a separate
address.
This version also supports multiple external networks - ropsten,xdai -
as well as local ganache (running on the same host as the browser)